Rewrite extract to use Babel bits
This rewrites extract to use Babel's Catalog to generate .pot
files. Two nice things come of this:

1. .pot files now have variable format flags (yay!)
2. less code to maintain

Additionally, we added an extractor that monkeypatches i18n and then
calls the Jinja2 extractor. That makes it possible for someone to use
pybabel extract rather than Puente extract if they so desired.