Fix gettext alias, jingo test project settings and add alias for 18n
This fixes the gettext alias so that it has the correct number of
arguments.

This undoes 3264309 which changed the jingo test project settings to
not add 'jingo.ext.i18n' to the beginning of the extensions list with
the theory that we don't need it since the
'puente.ext.PuenteI18nExtension' stomps on it later. However, it turns
out Jingo explicitly looks for 'jinja2.ext.i18n' in the extensions
list and does stuff based on that. Awesome. So this readds that line
with a note about why it's there.

Further, I got tired of typing PuenteI18nExtension and added an i18n
alias just like Jinja2 has.

Fixes #35