Remove the "make gettext output safe" code
Jinja2's newstyle gettext coupled with autoescape does the right thing
and we don't need to additionally do anything with gettext and
ngettext. Given that, we can nix all that code and related expository
text.

Further, I rewrote the infrastructure for the relevant tests so that
it's doing less mocking and now uses the Django translation stuff.

I added some new tests so we're covering all the combinations of
things so that we can be sure that Puente is doing the right thing.

I also updated the Migrating from Tower guide to cover changing
autoescape (it was suggested you set it to False in the jingo docs
(yeesh!)) and newstyle (Tower set up oldstyle gettext).

Fixes #39